Fixed default flag value error.
Allow Nil to be used as a default Flag value, but raise an error when
that Flag is added to a Flags collection since Nil isn't an option.

We've previously used `@default = uninitialized Types` instead of 
simply `@default = nil`, which is considered unsafe and seems to fail
in newer versions of Crystal. For this reason we'll use Nil as an
invalid default type value.